Refactoring: Move classes to separate files
